refactor(adapters): one SQL query compilation, two dialects
SQLite and Postgres each carried a near-line-identical copy of the
query walk — where clauses, the null-policy predicate, the operator
map, the comma-separated `in` convention, the identity ORDER BY
fallback, limit/offset. adapters/driven/sql_query_builder.rb now spells
it once; each adapter supplies only what a dialect genuinely differs
on (placeholders, contains/empty-in rendering, numeric casting, column
vs jsonb-path expressions, null ordering, execution). Both adapters
also move outside their folders — sqlite.rb splits its DDL into
sqlite/schema_builder.rb, its column codec into sqlite/codec.rb, and
SqliteProjection into sqlite/projection.rb. Verified against the full
parity suite, Postgres era dance included, before this commit.
