+++ lib/hecksagain/adapters/driven/sqlite.adapter
